requirements.txt
streamlit>=1.31
google-generativeai>=0.3.2

psycopg2-binary>=2.9.9
python-dotenv>=1.0.1

pandas>=2.1
numpy>=1.26

tqdm>=4.66
